+++ extensions/apple-reminders/src/hooks/useViewReminders.ts
@@ -1,10 +1,12 @@
 import { Icon, Image } from "@raycast/api";
 import { useCachedState } from "@raycast/utils";
-import { isBefore, isThisWeek, isToday, isTomorrow, parseISO, startOfDay } from "date-fns";
+import { isBefore, isSameDay, parseISO } from "date-fns";
 import { compareAsc } from "date-fns";
 import { partition } from "lodash";
 import { useMemo } from "react";
 import React from "react";
+
+import { displayDueDate, isOverdue, parseDay } from "../helpers";
 
 import { Data, Priority, Reminder } from "./useData";
 
@@ -89,43 +91,35 @@ export const groupByOptions: GroupByOptions = [
 ];
 
 export function groupByDueDates(reminders: Reminder[]) {
-  const overdue: Reminder[] = [];
-  const today: Reminder[] = [];
-  const tomorrow: Reminder[] = [];
-  const upcoming: Reminder[] = [];
-  const other: Reminder[] = [];
+  const [dated, notDated] = partition(reminders, (reminder: Reminder) => reminder.dueDate !== null);
+  const [overdue, upcoming] = partition(dated, (reminder: Reminder) => reminder.dueDate && isOverdue(reminder.dueDate));
+  const allDueDates = [...new Set(upcoming.map((reminder) => parseDay(reminder.dueDate).toISOString()))];
+  allDueDates.sort();
 
-  reminders.forEach((reminder) => {
-    if (!reminder.dueDate) {
-      other.push(reminder);
-    } else {
-      const dueDate = parseISO(reminder.dueDate);
-
-      if (isBefore(dueDate, startOfDay(new Date()))) {
-        overdue.push(reminder);
-      } else if (isToday(dueDate)) {
-        today.push(reminder);
-      } else if (isTomorrow(dueDate)) {
-        tomorrow.push(reminder);
-      } else if (isThisWeek(dueDate, { weekStartsOn: 1 })) {
-        upcoming.push(reminder);
-      } else {
-        other.push(reminder);
-      }
-    }
+  const sections = allDueDates.map((date) => {
+    const remindersOnDate = upcoming.filter((reminder) => parseDay(reminder.dueDate).toISOString() === date);
+    return {
+      title: displayDueDate(date),
+      reminders: remindersOnDate,
+    };
   });
 
-  const sections = [
-    { title: "Overdue", reminders: overdue },
-    { title: "Today", reminders: today },
-    { title: "Tomorrow", reminders: tomorrow },
-    { title: "Upcoming", reminders: upcoming },
-    { title: "Other", reminders: other },
-  ];
+  if (overdue.length > 0) {
+    sections.unshift({
+      title: "Overdue",
+      reminders: overdue,
+    });
+  }
+
+  if (notDated.length > 0) {
+    sections.push({
+      title: "No due date",
+      reminders: notDated,
+    });
+  }
 
   return sections;
 }
-
 export function groupByPriorities(reminders: Reminder[]) {
   const priorities: { name: string; value: Priority }[] = [
     { name: "High", value: "high" },
@@ -172,19 +166,25 @@ export default function useViewReminders(listId: string, { data }: { data?: Data
     `show-completed-reminders-${listId}`,
     false,
   );
-  const [sortBy, setSortBy] = useCachedState<SortByOption>(`sort-by-${listId}`, "default");
-  const [groupBy, setGroupBy] = useCachedState<GroupByOption>(`group-by-${listId}`, "default");
+
+  const viewDefault = listId === "today" || listId === "scheduled" ? "dueDate" : "default";
+
+  const [sortBy, setSortBy] = useCachedState<SortByOption>(`sort-by-${listId}`, viewDefault);
+  const [groupBy, setGroupBy] = useCachedState<GroupByOption>(`group-by-${listId}`, viewDefault);
   const [orderBy, setOrderBy] = useCachedState<OrderByOption>(`order-by-${listId}`, "asc");
 
   const reminders = useMemo(() => data?.reminders ?? [], [data]);
 
   const filteredReminders = useMemo(() => {
-    return (
-      reminders.filter((reminder) => {
-        if (listId === "all") return true;
-        return reminder.list?.id === listId;
-      }) ?? []
-    );
+    return reminders.filter((reminder) => {
+      if (listId === "all") return true;
+      if (listId === "today")
+        return reminder.dueDate
+          ? isOverdue(reminder.dueDate) || isSameDay(new Date(reminder.dueDate), parseDay())
+          : false;
+      if (listId === "scheduled") return !!reminder.dueDate;
+      return reminder.list?.id === listId;
+    });
   }, [listId, reminders]);
 
   const { sortByProp, sortedReminders, orderByProp } = useMemo(() => {
@@ -248,7 +248,17 @@ export default function useViewReminders(listId: string, { data }: { data?: Data
 
     switch (groupBy) {
       case "default": {
-        const title = listId === "all" ? "All" : data?.lists.find((list) => list.id === listId)?.title ?? "Reminders";
+        let title = "Reminders";
+        if (listId === "all") {
+          title = "All";
+        } else if (listId === "today") {
+          title = "Today";
+        } else if (listId === "scheduled") {
+          title = "Scheduled";
+        } else {
+          title = data?.lists.find((list) => list.id === listId)?.title ?? "Reminders";
+        }
+
         sections = [{ title, reminders: incompleteReminders }];
         break;
       }