+++ extensions/lastpass/src/cli.ts
@@ -26,12 +26,9 @@ const serializeFromJson = (jsonArray: string): Account[] => {
   return res;
 };
 
-const mask = (it: string, pattern: RegExp) => it.replace(pattern, "*".repeat(pattern.toString().length));
-
-const execute = async (command: string, opts?: { maskPattern?: RegExp }) => {
+const execute = async (command: string) => {
   const PATH = "/usr/gnu/bin:/usr/local/bin:/bin:/usr/bin:.:/opt/homebrew/bin";
   const wrappedCommand = `zsh -l -c 'export PATH="$PATH:${PATH}" && ${command}'`;
-  const maskPattern = opts?.maskPattern || new RegExp("");
 
   console.log(`Executing: ${wrappedCommand}`);
   const startTimestamp = Date.now();
@@ -43,7 +40,7 @@ const execute = async (command: string, opts?: { maskPattern?: RegExp }) => {
         console.error(`[${tookSeconds}s] Failed:\n${stderr}`);
         rej({
           ...error,
-          message: mask(error.message, maskPattern),
+          message: error.message,
         });
       }
       console.log(`[${tookSeconds}s] Success:\n${stdout}`);
@@ -52,54 +49,46 @@ const execute = async (command: string, opts?: { maskPattern?: RegExp }) => {
   );
 };
 
-const formatCommand = (subcommand: string, args: { password: string; maskPattern?: RegExp }) => {
-  const { password, maskPattern } = { maskPattern: new RegExp(""), ...args };
+const authorize = (subcommand: string, opts: { password: string }) => {
+  const { password } = opts;
   const quote = password.includes('"') ? "'" : '"';
-  const command = `echo ${quote}${password}${quote} | LPASS_DISABLE_PINENTRY=1 lpass ${subcommand}`;
-  return mask(command, maskPattern);
+  const maskedCommand = `echo ${quote}${password}${quote} | LPASS_DISABLE_PINENTRY=1 lpass ${subcommand}`;
+  return maskedCommand;
 };
 
 export const lastPass = (email: string, password: string) => {
-  const maskPattern = new RegExp(password);
-
   return {
     isLogged: () =>
-      execute(formatCommand("status", { password, maskPattern }), {})
+      execute(authorize("status", { password }))
         .then((stdout) => stdout.includes(email))
         .catch(() => false),
 
-    login: () => execute(formatCommand(`login ${email}`, { password, maskPattern }), { maskPattern }),
-
-    // returns exit code > 0 if failed, so just need to convert return type to Promise<void>
-    sync: () =>
-      execute("sync").then(() => {
-        /* noop */
-      }),
+    login: () => execute(authorize(`login ${email}`, { password })),
 
     show: (id: string, opts: { sync: "auto" | "now" | "no" } = { sync: "auto" }): Promise<Account> =>
-      execute(formatCommand(`show --sync=${opts.sync} --json ${id}`, { password, maskPattern }), { maskPattern }).then(
+      execute(authorize(`show --sync=${opts.sync} --json ${id}`, { password })).then(
         (stdout) => serializeFromJson(stdout)[0]
       ),
 
     list: (opts: { sync: "auto" | "now" | "no" } = { sync: "auto" }) =>
-      execute(
-        formatCommand(`ls --sync=${opts.sync} --format="%ai<=>%an<=>%au<=>%ap<=>%al"`, { password, maskPattern }),
-        { maskPattern }
-      ).then((stdout) => {
-        const items: { id: string; name: string; username: string; password: string; url: string }[] = stdout
-          .split("\n")
-          .map((line) => {
-            const [id, name, username, password, url] = line.split("<=>");
-            return { id, name, username, password, url };
-          })
-          .filter(({ name }) => name);
-        return items;
-      }),
+      execute(authorize(`ls --sync=${opts.sync} --format="%ai<=>%an<=>%au<=>%ap<=>%al"`, { password })).then(
+        (stdout) => {
+          const items: { id: string; name: string; username: string; password: string; url: string }[] = stdout
+            .split("\n")
+            .map((line) => {
+              const [id, name, username, password, url] = line.split("<=>");
+              return { id, name, username, password, url };
+            })
+            .filter(({ name }) => name);
+          return items;
+        }
+      ),
 
     export: (opts: { sync: "auto" | "now" | "no" } = { sync: "auto" }) =>
       execute(
-        formatCommand(`export --sync=${opts.sync} --fields=id,name,username,password,url`, { password, maskPattern }),
-        { maskPattern }
+        authorize(`export --sync=${opts.sync} --fields=id,name,username,password,url`, {
+          password,
+        })
       ).then((stdout) => {
         const items: { id: string; name: string; username: string; password: string; url: string }[] = stdout
           .split("\n")