+++ extensions/reclaim-ai/src/hooks/useEvent.tsx
@@ -2,16 +2,14 @@ import { Icon, getPreferenceValues, open } from "@raycast/api";
 import { format, isWithinInterval } from "date-fns";
 import { useCallback } from "react";
 import { Event } from "../types/event";
+import { NativePreferences } from "../types/preferences";
 import { axiosPromiseData } from "../utils/axiosPromise";
 import { formatDisplayEventHours, formatDisplayHours } from "../utils/dates";
 import { parseEmojiField } from "../utils/string";
 import reclaimApi from "./useApi";
 import { ApiResponseEvents, EventActions } from "./useEvent.types";
-import { useUser } from "./useUser";
 import { useTask } from "./useTask";
-import { NativePreferences } from "../types/preferences";
-import { CalendarAccount } from "../types/account";
-import { filterMultipleOutDuplicateEvents } from "../utils/events";
+import { useUser } from "./useUser";
 
 const useEvent = () => {
   const { fetcher } = reclaimApi();
@@ -24,23 +22,13 @@ const useEvent = () => {
       const strStart = format(start, "yyyy-MM-dd");
       const strEnd = format(end, "yyyy-MM-dd");
 
-      const [accountsResponse, accountsError] = await axiosPromiseData<CalendarAccount[]>(
-        fetcher("/accounts", {
-          method: "GET",
-        })
-      );
-
-      if (!accountsResponse || accountsError) throw accountsError;
-
       const [eventsResponse, error] = await axiosPromiseData<ApiResponseEvents>(
         fetcher("/events?sourceDetails=true", {
           method: "GET",
           params: {
             start: strStart,
             end: strEnd,
-            calendarIds: accountsResponse
-              .flatMap(({ connectedCalendars }) => connectedCalendars.map(({ id }) => id))
-              .join(","),
+            allConnected: true,
           },
         })
       );
@@ -48,7 +36,7 @@ const useEvent = () => {
       if (!eventsResponse || error) throw error;
 
       // Filter out events that are synced, managed by Reclaim and part of multiple calendars
-      return filterMultipleOutDuplicateEvents(eventsResponse);
+      return eventsResponse;
     } catch (error) {
       console.error("Error while fetching events", error);
     }