+++ extensions/teleport/assets/scripts/teleport-application-list.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+tsh apps ls --format=json