+++ extensions/apple-reminders/CHANGELOG.md
@@ -1,3 +1,12 @@
 # Apple Reminders Changelog
 
-## [Initial Version] - 2023-06-27
+## [Today and Scheduled views] - 2023-11-03
+
+- Add support for today and scheduled views, similar to those in the native Reminders app.
+
+## [Recurring remidners and better group by dates] - 2023-10-24
+
+- Add support for basic recurring reminders in `Create Reminder` and `Quick Add Reminder`
+- Improve "Group by due dates" behavior. Now, dates are grouped based on their actual due dates, regardless of whether they are due tomorrow, in 3 months, or in a year.
+
+## [Initial Version] - 2023-10-19