+++ extensions/mynaui-icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # MynaUI Icons Changelog
 
+## [Update Icons Version] - 2023-11-01
+
+- Updated mynaui-icons package to the latest version
+- Increased icons columns from 6 to 7
+- Added more information to README
+
 ## [Initial Version] - 2023-10-06
 
 Initial Release of mynaui-icons.